# refactor(get_environment_from_cache): Optimize SQL  query