--- a/api/environments/models.py
+++ b/api/environments/models.py
@@ -203,13 +203,13 @@ class Environment(
                     "heap_config",
                     "dynatrace_config",
                 )
-                environment = (
-                    cls.objects.select_related(*select_related_args)
-                    .filter(Q(api_key=api_key) | Q(api_keys__key=api_key))
-                    .distinct()
-                    .defer("description")
-                    .get()
+                base_qs = cls.objects.select_related(*select_related_args).defer(
+                    "description"
                 )
+                qs_one = base_qs.filter(api_key=api_key)
+                qs_two = base_qs.filter(api_keys__key=api_key)
+
+                environment = qs_one.union(qs_two).get()
                 environment_cache.set(
                     api_key, environment, timeout=settings.ENVIRONMENT_CACHE_SECONDS
                 )