Thanks for submitting a PR! Please check the boxes below:

- [x] I have run [`pre-commit`](https://docs.flagsmith.com/platform/contributing#pre-commit) to check linting
- [x] I have filled in the "Changes" section below?
- [x] I have filled in the "How did you test this code" section below?
- [x] I have used a [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) title for this Pull Request

## Changes

Use union to utitlize index instead of doing a sequential scan 

Old SQL query:

```
SELECT 
  DISTINCT "environments_environment"."id", 
  "environments_environment"."deleted_at", 
  "environments_environment"."name", 
  "environments_environment"."created_date", 
  "environments_environment"."project_id", 
  "environments_environment"."api_key", 
  "environments_environment"."minimum_change_request_approvals", 
  "environments_environment"."webhooks_enabled", 
  "environments_environment"."webhook_url", 
  "environments_environment"."allow_client_traits", 
  "environments_environment"."updated_at", 
  "environments_environment"."banner_text", 
  "environments_environment"."banner_colour", 
  "environments_environment"."hide_disabled_flags", 
  "environments_environment"."use_mv_v2_evaluation", 
  "environments_environment"."hide_sensitive_data", 
  "projects_project"."id", 
  "projects_project"."deleted_at", 
  "projects_project"."uuid", 
  "projects_project"."name", 
  "projects_project"."created_date", 
  "projects_project"."organisation_id", 
  "projects_project"."hide_disabled_flags", 
  "projects_project"."enable_dynamo_db", 
  "projects_project"."prevent_flag_defaults", 
  "projects_project"."enable_realtime_updates", 
  "projects_project"."only_allow_lower_case_feature_names", 
  "projects_project"."feature_name_regex", 
  "projects_project"."max_segments_allowed", 
  "projects_project"."max_features_allowed", 
  "projects_project"."max_segment_overrides_allowed", 
  "organisations_organisation"."id", 
  "organisations_organisation"."deleted_at", 
  "organisations_organisation"."uuid", 
  "organisations_organisation"."name", 
  "organisations_organisation"."has_requested_features", 
  "organisations_organisation"."webhook_notification_email", 
  "organisations_organisation"."created_date", 
  "organisations_organisation"."alerted_over_plan_limit", 
  "organisations_organisation"."stop_serving_flags", 
  "organisations_organisation"."restrict_project_create_to_admin", 
  "organisations_organisation"."persist_trait_data", 
  "organisations_organisation"."block_access_to_admin", 
  "organisations_organisation"."feature_analytics", 
  "amplitude_amplitudeconfiguration"."id", 
  "amplitude_amplitudeconfiguration"."deleted_at", 
  "amplitude_amplitudeconfiguration"."uuid", 
  "amplitude_amplitudeconfiguration"."api_key", 
  "amplitude_amplitudeconfiguration"."base_url", 
  "amplitude_amplitudeconfiguration"."environment_id", 
  "segment_segmentconfiguration"."id", 
  "segment_segmentconfiguration"."deleted_at", 
  "segment_segmentconfiguration"."uuid", 
  "segment_segmentconfiguration"."base_url", 
  "segment_segmentconfiguration"."api_key", 
  "segment_segmentconfiguration"."environment_id", 
  "heap_heapconfiguration"."id", 
  "heap_heapconfiguration"."deleted_at", 
  "heap_heapconfiguration"."uuid", 
  "heap_heapconfiguration"."base_url", 
  "heap_heapconfiguration"."api_key", 
  "heap_heapconfiguration"."environment_id", 
  "mixpanel_mixpanelconfiguration"."id", 
  "mixpanel_mixpanelconfiguration"."deleted_at", 
  "mixpanel_mixpanelconfiguration"."uuid", 
  "mixpanel_mixpanelconfiguration"."base_url", 
  "mixpanel_mixpanelconfiguration"."api_key", 
  "mixpanel_mixpanelconfiguration"."environment_id", 
  "dynatrace_dynatraceconfiguration"."id", 
  "dynatrace_dynatraceconfiguration"."deleted_at", 
  "dynatrace_dynatraceconfiguration"."uuid", 
  "dynatrace_dynatraceconfiguration"."base_url", 
  "dynatrace_dynatraceconfiguration"."api_key", 
  "dynatrace_dynatraceconfiguration"."environment_id", 
  "dynatrace_dynatraceconfiguration"."entity_selector" 
FROM 
  "environments_environment" 
  LEFT OUTER JOIN "environments_environmentapikey" ON (
    "environments_environment"."id" = "environments_environmentapikey"."environment_id"
  ) 
  INNER JOIN "projects_project" ON (
    "environments_environment"."project_id" = "projects_project"."id"
  ) 
  INNER JOIN "organisations_organisation" ON (
    "projects_project"."organisation_id" = "organisations_organisation"."id"
  ) 
  LEFT OUTER JOIN "amplitude_amplitudeconfiguration" ON (
    "environments_environment"."id" = "amplitude_amplitudeconfiguration"."environment_id"
  ) 
  LEFT OUTER JOIN "segment_segmentconfiguration" ON (
    "environments_environment"."id" = "segment_segmentconfiguration"."environment_id"
  ) 
  LEFT OUTER JOIN "heap_heapconfiguration" ON (
    "environments_environment"."id" = "heap_heapconfiguration"."environment_id"
  ) 
  LEFT OUTER JOIN "mixpanel_mixpanelconfiguration" ON (
    "environments_environment"."id" = "mixpanel_mixpanelconfiguration"."environment_id"
  ) 
  LEFT OUTER JOIN "dynatrace_dynatraceconfiguration" ON (
    "environments_environment"."id" = "dynatrace_dynatraceconfiguration"."environment_id"
  ) 
WHERE 
  (
    "environments_environment"."deleted_at" IS NULL 
    AND (
      "environments_environment"."api_key" = '<api_Key>' 
      OR "environments_environmentapikey"."key" = '<api_Key>'
    )
  ) 
LIMIT 
  21

```

Generated plan:

```
Limit  (cost=1050.42..1054.31 rows=21 width=2552) (actual time=15.146..15.149 rows=0 loops=1)
   Buffers: shared hit=8084
   ->  Unique  (cost=1050.42..1054.68 rows=23 width=2552) (actual time=15.145..15.148 rows=0 loops=1)
         Buffers: shared hit=8084
         ->  Sort  (cost=1050.42..1050.48 rows=23 width=2552) (actual time=15.144..15.147 rows=0 loops=1)
               Sort Key: environments_environment.id, environments_environment.deleted_at, environments_environment.name, environments_environment.created_date, environments_environment.project_id, environments_environment.api_key, environments_environment.minimum_change_request_approvals, environments_environment.
webhooks_enabled, environments_environment.webhook_url, environments_environment.allow_client_traits, environments_environment.updated_at, environments_environment.banner_text, environments_environment.banner_colour, environments_environment.hide_disabled_flags, environments_environment.use_mv_v2_evaluation, enviro
nments_environment.hide_sensitive_data, projects_project.deleted_at, projects_project.uuid, projects_project.name, projects_project.created_date, projects_project.organisation_id, projects_project.hide_disabled_flags, projects_project.enable_dynamo_db, projects_project.prevent_flag_defaults, projects_project.enable
_realtime_updates, projects_project.only_allow_lower_case_feature_names, projects_project.feature_name_regex, projects_project.max_segments_allowed, projects_project.max_features_allowed, projects_project.max_segment_overrides_allowed, organisations_organisation.deleted_at, organisations_organisation.uuid, organisa
tions_organisation.name, organisations_organisation.has_requested_features, organisations_organisation.webhook_notification_email, organisations_organisation.created_date, organisations_organisation.alerted_over_plan_limit, organisations_organisation.stop_serving_flags, organisations_organisation.restrict_project_c
reate_to_admin, organisations_organisation.persist_trait_data, organisations_organisation.block_access_to_admin, organisations_organisation.feature_analytics, amplitude_amplitudeconfiguration.id, amplitude_amplitudeconfiguration.deleted_at, amplitude_amplitudeconfiguration.uuid, amplitude_amplitudeconfiguration.api
_key, amplitude_amplitudeconfiguration.base_url, amplitude_amplitudeconfiguration.environment_id, segment_segmentconfiguration.id, segment_segmentconfiguration.deleted_at, segment_segmentconfiguration.uuid, segment_segmentconfiguration.base_url, segment_segmentconfiguration.api_key, segment_segmentconfiguration.env
ironment_id, heap_heapconfiguration.id, heap_heapconfiguration.deleted_at, heap_heapconfiguration.uuid, heap_heapconfiguration.base_url, heap_heapconfiguration.api_key, heap_heapconfiguration.environment_id, mixpanel_mixpanelconfiguration.id, mixpanel_mixpanelconfiguration.deleted_at, mixpanel_mixpanelconfiguration
.uuid, mixpanel_mixpanelconfiguration.base_url, mixpanel_mixpanelconfiguration.api_key, mixpanel_mixpanelconfiguration.environment_id, dynatrace_dynatraceconfiguration.id, dynatrace_dynatraceconfiguration.deleted_at, dynatrace_dynatraceconfiguration.uuid, dynatrace_dynatraceconfiguration.base_url, dynatrace_dynatra
ceconfiguration.api_key, dynatrace_dynatraceconfiguration.environment_id, dynatrace_dynatraceconfiguration.entity_selector
               Sort Method: quicksort  Memory: 25kB
               Buffers: shared hit=8084
               ->  Nested Loop Left Join  (cost=1.42..1049.90 rows=23 width=2552) (actual time=15.108..15.112 rows=0 loops=1)
                     Join Filter: (environments_environment.id = dynatrace_dynatraceconfiguration.environment_id)
                     Buffers: shared hit=8084
                     ->  Nested Loop Left Join  (cost=1.42..1046.76 rows=23 width=1368) (actual time=15.108..15.111 rows=0 loops=1)
                           Join Filter: (environments_environment.id = mixpanel_mixpanelconfiguration.environment_id)
                           Buffers: shared hit=8084
                           ->  Nested Loop Left Join  (cost=1.42..1036.89 rows=23 width=1306) (actual time=15.107..15.110 rows=0 loops=1)
                                 Join Filter: (environments_environment.id = heap_heapconfiguration.environment_id)
                                 Buffers: shared hit=8084
                                 ->  Nested Loop Left Join  (cost=1.42..1026.96 rows=23 width=844) (actual time=15.107..15.109 rows=0 loops=1)
                                       Buffers: shared hit=8084
                                       ->  Nested Loop Left Join  (cost=1.27..1023.33 rows=23 width=778) (actual time=15.106..15.108 rows=0 loops=1)
                                             Buffers: shared hit=8084
                                             ->  Nested Loop  (cost=1.14..1019.70 rows=23 width=691) (actual time=15.105..15.107 rows=0 loops=1)
                                                   Buffers: shared hit=8084
                                                   ->  Nested Loop  (cost=0.85..1012.31 rows=23 width=608) (actual time=15.105..15.106 rows=0 loops=1)
                                                         Buffers: shared hit=8084
                                                         ->  Merge Left Join  (cost=0.56..1005.16 rows=23 width=524) (actual time=15.104..15.105 rows=0 loops=1)
                                                               Merge Cond: (environments_environment.id = environments_environmentapikey.environment_id)
                                                               Filter: (((environments_environment.api_key)::text = '<api_key>'::text) OR ((environments_environmentapikey.key)::text = '<api_key>'::text))
                                                               Rows Removed by Filter: 23294
                                                               Buffers: shared hit=8084
                                                               ->  Index Scan using environments_environment_pkey on environments_environment  (cost=0.29..896.12 rows=23153 width=524) (actual time=0.045..11.901 rows=23180 loops=1)
                                                                     Filter: (deleted_at IS NULL)
                                                                     Rows Removed by Filter: 3881
                                                                     Buffers: shared hit=7656
                                                               ->  Index Scan using environments_environmentapikey_environment_id_bae78ba1 on environments_environmentapikey  (cost=0.28..35.10 rows=1046 width=31) (actual time=0.024..0.370 rows=1075 loops=1)
                                                                     Buffers: shared hit=428
                                                         ->  Index Scan using projects_project_pkey on projects_project  (cost=0.29..0.31 rows=1 width=84) (never executed)
                                                               Index Cond: (id = environments_environment.project_id)
                                                   ->  Index Scan using organisations_organisation_pkey on organisations_organisation  (cost=0.29..0.32 rows=1 width=83) (never executed)
                                                         Index Cond: (id = projects_project.organisation_id)
                                             ->  Index Scan using amplitude_amplitudeconfiguration_environment_id_key on amplitude_amplitudeconfiguration  (cost=0.14..0.16 rows=1 width=87) (never executed)
                                                   Index Cond: (environments_environment.id = environment_id)
                                       ->  Index Scan using segment_segmentconfiguration_environment_id_key on segment_segmentconfiguration  (cost=0.14..0.16 rows=1 width=66) (never executed)
                                             Index Cond: (environments_environment.id = environment_id)
                                 ->  Materialize  (cost=0.00..1.38 rows=25 width=462) (never executed)
                                       ->  Seq Scan on heap_heapconfiguration  (cost=0.00..1.25 rows=25 width=462) (never executed)
                           ->  Materialize  (cost=0.00..2.33 rows=22 width=62) (never executed)
                                 ->  Seq Scan on mixpanel_mixpanelconfiguration  (cost=0.00..2.22 rows=22 width=62) (never executed)
                     ->  Materialize  (cost=0.00..1.09 rows=6 width=1184) (never executed)
                           ->  Seq Scan on dynatrace_dynatraceconfiguration  (cost=0.00..1.06 rows=6 width=1184) (never executed)
 Planning Time: 8.182 ms
 Execution Time: 15.534 ms


```
Query generated by this PR
```
(
  SELECT 
    "environments_environment"."id", 
    "environments_environment"."deleted_at", 
    "environments_environment"."name", 
    "environments_environment"."created_date", 
    "environments_environment"."project_id", 
    "environments_environment"."api_key", 
    "environments_environment"."minimum_change_request_approvals", 
    "environments_environment"."webhooks_enabled", 
    "environments_environment"."webhook_url", 
    "environments_environment"."allow_client_traits", 
    "environments_environment"."updated_at", 
    "environments_environment"."banner_text", 
    "environments_environment"."banner_colour", 
    "environments_environment"."hide_disabled_flags", 
    "environments_environment"."use_mv_v2_evaluation", 
    "environments_environment"."hide_sensitive_data", 
    "projects_project"."id", 
    "projects_project"."deleted_at", 
    "projects_project"."uuid", 
    "projects_project"."name", 
    "projects_project"."created_date", 
    "projects_project"."organisation_id", 
    "projects_project"."hide_disabled_flags", 
    "projects_project"."enable_dynamo_db", 
    "projects_project"."prevent_flag_defaults", 
    "projects_project"."enable_realtime_updates", 
    "projects_project"."only_allow_lower_case_feature_names", 
    "projects_project"."feature_name_regex", 
    "projects_project"."max_segments_allowed", 
    "projects_project"."max_features_allowed", 
    "projects_project"."max_segment_overrides_allowed", 
    "organisations_organisation"."id", 
    "organisations_organisation"."deleted_at", 
    "organisations_organisation"."uuid", 
    "organisations_organisation"."name", 
    "organisations_organisation"."has_requested_features", 
    "organisations_organisation"."webhook_notification_email", 
    "organisations_organisation"."created_date", 
    "organisations_organisation"."alerted_over_plan_limit", 
    "organisations_organisation"."stop_serving_flags", 
    "organisations_organisation"."restrict_project_create_to_admin", 
    "organisations_organisation"."persist_trait_data", 
    "organisations_organisation"."block_access_to_admin", 
    "organisations_organisation"."feature_analytics", 
    "amplitude_amplitudeconfiguration"."id", 
    "amplitude_amplitudeconfiguration"."deleted_at", 
    "amplitude_amplitudeconfiguration"."uuid", 
    "amplitude_amplitudeconfiguration"."api_key", 
    "amplitude_amplitudeconfiguration"."base_url", 
    "amplitude_amplitudeconfiguration"."environment_id", 
    "segment_segmentconfiguration"."id", 
    "segment_segmentconfiguration"."deleted_at", 
    "segment_segmentconfiguration"."uuid", 
    "segment_segmentconfiguration"."base_url", 
    "segment_segmentconfiguration"."api_key", 
    "segment_segmentconfiguration"."environment_id", 
    "heap_heapconfiguration"."id", 
    "heap_heapconfiguration"."deleted_at", 
    "heap_heapconfiguration"."uuid", 
    "heap_heapconfiguration"."base_url", 
    "heap_heapconfiguration"."api_key", 
    "heap_heapconfiguration"."environment_id", 
    "mixpanel_mixpanelconfiguration"."id", 
    "mixpanel_mixpanelconfiguration"."deleted_at", 
    "mixpanel_mixpanelconfiguration"."uuid", 
    "mixpanel_mixpanelconfiguration"."base_url", 
    "mixpanel_mixpanelconfiguration"."api_key", 
    "mixpanel_mixpanelconfiguration"."environment_id", 
    "dynatrace_dynatraceconfiguration"."id", 
    "dynatrace_dynatraceconfiguration"."deleted_at", 
    "dynatrace_dynatraceconfiguration"."uuid", 
    "dynatrace_dynatraceconfiguration"."base_url", 
    "dynatrace_dynatraceconfiguration"."api_key", 
    "dynatrace_dynatraceconfiguration"."environment_id", 
    "dynatrace_dynatraceconfiguration"."entity_selector" 
  FROM 
    "environments_environment" 
    INNER JOIN "projects_project" ON (
      "environments_environment"."project_id" = "projects_project"."id"
    ) 
    INNER JOIN "organisations_organisation" ON (
      "projects_project"."organisation_id" = "organisations_organisation"."id"
    ) 
    LEFT OUTER JOIN "amplitude_amplitudeconfiguration" ON (
      "environments_environment"."id" = "amplitude_amplitudeconfiguration"."environment_id"
    ) 
    LEFT OUTER JOIN "segment_segmentconfiguration" ON (
      "environments_environment"."id" = "segment_segmentconfiguration"."environment_id"
    ) 
    LEFT OUTER JOIN "heap_heapconfiguration" ON (
      "environments_environment"."id" = "heap_heapconfiguration"."environment_id"
    ) 
    LEFT OUTER JOIN "mixpanel_mixpanelconfiguration" ON (
      "environments_environment"."id" = "mixpanel_mixpanelconfiguration"."environment_id"
    ) 
    LEFT OUTER JOIN "dynatrace_dynatraceconfiguration" ON (
      "environments_environment"."id" = "dynatrace_dynatraceconfiguration"."environment_id"
    ) 
  WHERE 
    (
      "environments_environment"."deleted_at" IS NULL 
      AND "environments_environment"."api_key" = '<api_key>'
    ) 
  ORDER BY 
    "environments_environment"."id" ASC
) 
UNION 
  (
    SELECT 
      "environments_environment"."id", 
      "environments_environment"."deleted_at", 
      "environments_environment"."name", 
      "environments_environment"."created_date", 
      "environments_environment"."project_id", 
      "environments_environment"."api_key", 
      "environments_environment"."minimum_change_request_approvals", 
      "environments_environment"."webhooks_enabled", 
      "environments_environment"."webhook_url", 
      "environments_environment"."allow_client_traits", 
…