@gagantrivedi 2023-10-03
> @gagantrivedi It looks to me that the issue is that for some reason the query is not being optimized to apply the below conditions during the JOIN and not after making the whole Cartesian Product, right? :
> 
> ```
>       "environments_environment"."api_key" = '<api_Key>' 
>       OR "environments_environmentapikey"."key" = '<api_Key>'
> ```
> 
> Should we find a way to force the use of the conditions in the JOIN's ON clause instead of trying to build a UNION by using for example `FilteredRelation`?

Sorry, I don't follow. The issue is that Postgres is not using the index because of OR