@novakzaballa 2023-10-03
> > @gagantrivedi It looks to me that the issue is that for some reason the query is not being optimized to apply the below conditions during the JOIN and not after making the whole Cartesian Product, right? :
> > ```
> >       "environments_environment"."api_key" = '<api_Key>' 
> >       OR "environments_environmentapikey"."key" = '<api_Key>'
> > ```
> > 
> > 
> >     
> >       
> >     
> > 
> >       
> >     
> > 
> >     
> >   
> > Should we find a way to force the use of the conditions in the JOIN's ON clause instead of trying to build a UNION by using for example `FilteredRelation`?
> 
> Sorry, I don't follow. The issue is that Postgres is not using the index because of OR

Well, I don't know the final objective of this query, but a first glance looks like the original query with the below modifications would probably be more efficient than using UNION, and should use the same indexes. We should be able to achive this by using `FilteredRelation` with Django I think:
```
...
  "dynatrace_dynatraceconfiguration"."environment_id", 
  "dynatrace_dynatraceconfiguration"."entity_selector" 
FROM 
  "environments_environment" 
  LEFT OUTER JOIN "environments_environmentapikey" ON (
    "environments_environment"."id" = "environments_environmentapikey"."environment_id" and "environments_environmentapikey"."key" = '<api_Key>' and "environments_environmentapikey"."key" <> "environments_environment"."api_key"
  ) 
  INNER JOIN "projects_project" ON (
    "environments_environment"."project_id" = "projects_project"."id"
  ) 
...
WHERE 
  (
    "environments_environment"."deleted_at" IS NULL 
    AND 
      "environments_environment"."api_key" = '<api_Key>' 
  ) 

```
Probably an even better solution would be avoiding joining with `environments_environmentapikey` at all, and instead use EXISTS to check if the <api_key> is there, something like:
```
...
FROM 
  "environments_environment" 
  INNER JOIN "projects_project" ON (
    "environments_environment"."project_id" = "projects_project"."id"
  ) 
...
  WHERE 
  (
    "environments_environment"."deleted_at" IS NULL
    AND (
      "environments_environment"."api_key" = '<api_Key>' 
      OR EXISTS (SELECT 1 FROM "environments_environmentapikey" WHERE "environments_environment"."id" = "environments_environmentapikey"."environment_id" and "environments_environmentapikey"."key" = '<api_Key>')
    )
  ) 
```
Also, I don't know why we are using the DISTINCT at the beginning of the query, but if it was only because of the duplicate items that joining with the `environments_environmentapikey` table could cause, any of these 2 options should allow us to get rid of the DISTINCT.

If you think it's worth it, could you please try these 2 options and see the performance in the DB, I don't think I have access.